Cleaning up the API
Files are split into:
EngineEventSubscriber, takes events from engine and sends them to the client.
SimulationRunner, handles simulation controls.
SimulationWebSocketService, handles the websocket and connections.
SnapshotHandler: Builds snapshots from the engine based on client requests.
