+++ API/Program.cs
@@ -7,9 +7,18 @@ using Engine.Cost;
 using API.EngineManager;
 using Protocol;
 using Google.Protobuf;
+using Microsoft.Extensions.Logging;
 
+/// <summary>
+/// Entry point for the SmartEV API application.
+/// </summary>
 public static class Program
 {
+    /// <summary>
+    /// The main entry point for the SmartEV API application.
+    /// </summary>
+    /// <param name="args">The command-line arguments.</param>
+    /// <returns>A <see cref="Task"/> representing the result of the asynchronous operation that initializes the application.</returns>
     public static async Task Main(string[] args)
     {
         var builder = WebApplication.CreateBuilder(args);
@@ -18,82 +27,87 @@ public static class Program
         builder.Services.AddCors(options =>
         {
             options.AddPolicy("AllowFrontend", policy =>
-            {
                 policy.WithOrigins("http://localhost:5173")
                       .AllowAnyMethod()
-                      .AllowAnyHeader();
-            });
+                      .AllowAnyHeader());
         });
 
         var app = builder.Build();
 
         if (app.Environment.IsDevelopment())
-        {
             app.UseDeveloperExceptionPage();
-        }
 
         app.UseWebSockets();
         app.UseCors("AllowFrontend");
 
-        app.MapSimulationWebSocket();
-
-        app.MapGet("/weights", async () =>
-        {
-            var weights = CostWeightMetadata.All;
-            return Results.Ok(weights);
-        });
+        app.MapGet("/weights", async () => Results.Ok(CostWeightMetadata.All));
 
         app.MapPost("/init-engine", async (EngineManager.EngineManager engineManager, EngineInitConfigDTO config) =>
         {
             var result = await engineManager.InitializeAsync(config, services =>
             {
                 services.AddLogging();
-                services.AddSingleton<SimulationChannel>();
-                services.AddSingleton<SimulationEngineService>();
-                services.AddSingleton<IEngineEventSubscriber>(sp => sp.GetRequiredService<SimulationEngineService>());
-                services.AddSingleton<EnvelopeWebSocketHandler>();
+                services.AddSingleton<IEngineEventSubscriber, EngineEventSubscriber>();
+                services.AddSingleton<SnapshotHandler>();
+                services.AddSingleton<SimulationWebSocketService>();
+                services.AddSingleton<IEventSender>(sp => sp.GetRequiredService<SimulationWebSocketService>());
+                services.AddSingleton<SimulationRunner>();
             });
 
+            if (!result)
+            {
+                return Results.BadRequest("Initialization failed");
+            }
+
             var stationService = engineManager.GetEngineService<StationService>();
-            var initData = new InitEngineData();
-
-            try
-            {
-                foreach (var station in stationService.GetAllStations())
-                {
-                    var stationInit = new StationInit
-                    {
-                        Id = station.Id,
-                        Address = station.Address,
-                        Pos = new Position { Lat = station.Position.Latitude, Lon = station.Position.Longitude },
-                    };
-                    initData.Stations.Add(stationInit);
-
-                    foreach (var charger in station.Chargers)
-                    {
-                        var chargerProto = new Charger
-                        {
-                            Id = charger.Id,
-                            MaxPowerKw = charger.MaxPowerKW,
-                            StationId = station.Id,
-                            IsDual = charger.GetSockets().Length > 1,
-                        };
-                        initData.Chargers.Add(chargerProto);
-                    }
-                }
-            }
-            catch
-            {
-            }
+            var initData = InitEngineDataBuilder.BuildInitEngineData(stationService);
 
             var envelope = new Envelope { InitEngineData = initData };
             var data = envelope.ToByteArray();
 
-            return result
-                ? Results.File(data, "application/octet-stream")
-                : Results.BadRequest("Initialization failed");
+            return Results.File(data, "application/octet-stream");
         });
 
-        await app.RunAsync();
+        app.Map("/ws/simulation", async context =>
+        {
+            if (!context.WebSockets.IsWebSocketRequest)
+            {
+                context.Response.StatusCode = StatusCodes.Status400BadRequest;
+                return;
+            }
+
+            var engineManager = context.RequestServices.GetRequiredService<EngineManager.EngineManager>();
+
+            if (!engineManager.IsInitialized)
+            {
+                context.Response.StatusCode = StatusCodes.Status503ServiceUnavailable;
+                await context.Response.WriteAsync("Engine not initialized. Call /init-engine first.");
+                return;
+            }
+
+            var webSocketService = engineManager.GetEngineService<SimulationWebSocketService>();
+            var simulationRunner = engineManager.GetEngineService<SimulationRunner>();
+
+            using var webSocket = await context.WebSockets.AcceptWebSocketAsync();
+
+            await simulationRunner.StartAsync();
+
+            try
+            {
+                await webSocketService.HandleConnectionAsync(webSocket, context.RequestAborted);
+            }
+            catch (Exception ex)
+            {
+                var loggerFactory = context.RequestServices.GetRequiredService<ILoggerFactory>();
+                var logger = loggerFactory.CreateLogger("WebSocket");
+                logger.LogError(ex, "WebSocket error");
+            }
+            finally
+            {
+                await simulationRunner.StopAsync();
+            }
+        });
+
+        app.Run();
     }
 }
